Collectd reporter supports disabled metric attributes.
